requirements.txt
langchain-groq
langchain-openai
langchain-tavily
langchain-community
python-dotenv
uvicorn
fastapi
pydantic
streamlit
langgraph
langchain-core
ReAct
langchain